Refactor: assemble code for foreign components into one place.
